server/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary  # Use binary for easier install
pydantic
pyjwt
passlib[bcrypt]
python-dotenv
python-jose[cryptography]
pydantic[email]
python-multipart